Restructure project into modules
