+++ simulation/__init__.py
@@ -1,0 +1,3 @@
+from .simulation import Simulation
+from .gas_phase_sim import GasPhaseSimulation
+from .solid_state_sim import SolidStateSimulation
